hotel_project/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.8
pillow==12.1.1
pycparser==3.0
PyJWT==2.12.1
PyMySQL==1.1.2
sqlparse==0.5.5
tzdata==2025.3
